Ray tracer implementation using Accelerate

+++ src/Animation.hs
@@ -13,7 +13,7 @@ import Linear ((*^))
 import GHC.Generics
 
 data Keyframe = Keyframe { camera :: CF.Camera
-                         , time :: Double }
+                         , time :: Float }
                          deriving (Generic)
 
 data InterpolationMethod = Linear
@@ -44,21 +44,21 @@ validateKeyframes frs = if time (head frs) == 0 && time (last frs) == 1
 
 generateFrames :: Animation -> [CF.Config]
 generateFrames animation = let
-    stepsize = (1 :: Double) / fromIntegral (nFrames animation - 1)
+    stepsize = (1 :: Float) / fromIntegral (nFrames animation - 1)
     -- Take the first keyframe from the scene in the config
     -- Also sort the frames by time
     frames = sortBy (comparing time) $ keyframes animation
     points = (* stepsize) . fromIntegral <$> [0 .. nFrames animation - 1]
     in map (makeFrame animation frames) points
 
-makeFrame :: Animation -> [Keyframe] -> Double -> CF.Config
+makeFrame :: Animation -> [Keyframe] -> Float -> CF.Config
 makeFrame animation frames point = let
         scn = scene animation
         mtd = interpolation animation
     in CF.Config { CF.camera = interpolate mtd frames point
                  , CF.scene = scn }
 
-interpolate :: InterpolationMethod -> [Keyframe] -> Double -> CF.Camera
+interpolate :: InterpolationMethod -> [Keyframe] -> Float -> CF.Camera
 interpolate method frames t = let
         findFrames (fr1 : fr2 : frs) = if t >= time fr1 && t < time fr2
             then (fr1, fr2)
@@ -68,7 +68,7 @@ interpolate method frames t = let
         (f1, f2) = findFrames frames
         t' = (t - time f1) / (time f2 - time f1)
 
-        f :: Fractional a => (Double -> a -> a) -> a -> a -> a
+        f :: Fractional a => (Float -> a -> a) -> a -> a -> a
         f = interpolationFunction method t'
 
         cam1 = camera f1
@@ -78,8 +78,8 @@ interpolate method frames t = let
                  , CF.lookAt = f (*^) (CF.lookAt cam1) (CF.lookAt cam2)
                  , CF.upVec = f (*^) (CF.upVec cam1) (CF.upVec cam2) }
 
-interpolationFunction :: Fractional a => InterpolationMethod -> Double
-                                      -> (Double -> a -> a)
+interpolationFunction :: Fractional a => InterpolationMethod -> Float
+                                      -> (Float -> a -> a)
                                       -> a -> a -> a
 {-# INLINE interpolationFunction #-}
 interpolationFunction method t times a b = case method of
